# Slave boomer does not hatch tasks as expected