--- a/.travis.yml
+++ b/.travis.yml
@@ -13,6 +13,7 @@ install:
   - go get github.com/zeromq/gomq
   - go get github.com/google/uuid
   - go get github.com/olekukonko/tablewriter
+  - go get github.com/stretchr/testify/assert
 
 script:
   - go test -timeout 1m -coverprofile=coverage.txt -covermode=atomic