--- a/runner_test.go
+++ b/runner_test.go
@@ -1,9 +1,12 @@
 package boomer
 
 import (
+	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
 )
 
 type HitOutput struct {
@@ -113,6 +116,75 @@ func TestSpawnWorkers(t *testing.T) {
 	currentClients := atomic.LoadInt32(&runner.numClients)
 	if currentClients > 3 {
 		t.Error("Spawning goroutines too fast, current count", currentClients)
+	}
+}
+
+func TestSpawnWorkersWithManyTasks(t *testing.T) {
+	var lock sync.Mutex
+	taskCalls := map[string]int{}
+
+	createTask := func(name string, weight int) *Task {
+		return &Task{
+			Name:   name,
+			Weight: weight,
+			Fn: func() {
+				lock.Lock()
+				defer lock.Unlock()
+				taskCalls[name] += 1
+			},
+		}
+	}
+	tasks := []*Task{
+		createTask(`one hundred`, 100),
+		createTask(`ten`, 10),
+		createTask(`one`, 1),
+	}
+
+	runner := newSlaveRunner("localhost", 5557, tasks, nil)
+	defer runner.close()
+
+	runner.client = newClient("localhost", 5557, runner.nodeID)
+
+	const numToSpawn int = 30
+	const hatchRate float64 = 10
+	runner.hatchRate = hatchRate
+
+	go runner.spawnWorkers(numToSpawn, runner.stopChan, runner.hatchComplete)
+	time.Sleep(4 * time.Second)
+
+	currentClients := atomic.LoadInt32(&runner.numClients)
+
+	assert.Equal(t, numToSpawn, int(currentClients))
+	lock.Lock()
+	hundreds := taskCalls[`one hundred`]
+	tens := taskCalls[`ten`]
+	ones := taskCalls[`one`]
+	lock.Unlock()
+
+	total := hundreds + tens + ones
+	t.Logf("total tasks run: %d\n", total)
+
+	assert.True(t, total > 111)
+
+	assert.True(t, ones > 1)
+	actPercentage := float64(ones) / float64(total)
+	expectedPercentage := 1.0 / 111.0
+	if actPercentage > 2*expectedPercentage || actPercentage < 0.5*expectedPercentage {
+		t.Errorf("Unexpected percentage of ones task: exp %v, act %v", expectedPercentage, actPercentage)
+	}
+
+	assert.True(t, tens > 10)
+	actPercentage = float64(tens) / float64(total)
+	expectedPercentage = 10.0 / 111.0
+	if actPercentage > 2*expectedPercentage || actPercentage < 0.5*expectedPercentage {
+		t.Errorf("Unexpected percentage of tens task: exp %v, act %v", expectedPercentage, actPercentage)
+	}
+
+	assert.True(t, hundreds > 100)
+	actPercentage = float64(hundreds) / float64(total)
+	expectedPercentage = 100.0 / 111.0
+	if actPercentage > 2*expectedPercentage || actPercentage < 0.5*expectedPercentage {
+		t.Errorf("Unexpected percentage of hundreds task: exp %v, act %v", expectedPercentage, actPercentage)
 	}
 }