### Problem
If I have a slice of tasks each with a given weight, I would expect the slave nodes to each spin up tasks with the relative weight of each task. However, because we _first_ loop through the tasks and then calc the `amount` for each task, there are some tasks that we spin up zero go routines for.

https://github.com/myzhan/boomer/blob/5424d6ac71e4c18b30dc87e5352c9f603abaab81/runner.go#L128-L165

Additionally, I notice `hatchCompleteFunc` might never be called if we send a signal on `quit` (but I don't _need_ that in this PR).

### Solution
Hatch goroutines at the same rate. However, choose the task randomly _after_ the goroutine has been hatched, not before.

@myzhan What would you suggest?