@myzhan 2019-12-18
> @myzhan it looks like the test failures aren't my fault, but preexisting?

Yes, there is a flaky test when running on Travis, I will find some times to figure out why. And thanks for your time, I will have a look at this today.