@myzhan 2019-12-21
@joshprzybyszewski-wf The weight of a boomer.Task was used to allocate goroutines to run the task. In the life cycle of a goroutine, it only run one task, which differs from how locustio do [here](https://github.com/locustio/locust/blob/master/locust/runners.py#L127). I think we can allocate at least one goroutine for a task, althought its weight is too low to get one, to solve the issue. What do you think? If you want to randomize task selection, you can put tasks in a [WeighingTaskSet](https://github.com/myzhan/boomer/blob/master/taskset.go#L25) to do the job.