@joshprzybyszewski-wf 2019-12-26
:TIL: about the `WeighingTaskSet` -- thank you! It does work, but it seems counter-intuitive to me to make consumers use a "task set" if they just have a bunch of tasks that they want to run.

Correct -- in master, each goroutine we spin up in a slave will only ever run one task. But is that desired? It limits the granularity of the percentage a slave node can run each task with (unless the developer uses a `WeighingTaskSet`). For example, if I want to run 5 clients on my slave node with three tasks where `A` is weighted at 45, `B` at 2, and `C` at 2, then `B` and `C` will _never_ get run because `A` is weighted 45/49. And 45/49 is greater than 4.5, which is the last round below 5 clients so it will receive all of the goroutines.

Instead, shouldn't we spin up the number of goroutines equal to the number of clients the slave node wants (as I've proposed in this PR)? Then each goroutine can pick from the random set of weighted tasks and execute each according to their weight. 

My perspective is that I would like to make this tool be as simple to understand/use as possible. Let me know what we should do to achieve that:)