@myzhan 2019-12-28
Yep, I agree with you on making it more simpler. I'm thinking about what will happen unexpectly if we randomly execute tasks according to their weights. What if Task A runs much slower than B and C?