@joshprzybyszewski-wf 2020-01-10
@myzhan sorry for the delay;

what will happen?

### On master commit (before this PR)

Assume 5 clients on slave node:

| task | weight | dedicated num client goroutines |
| - | - | - |
| A | 45 | 5 |
| B | 2 | 0 |
| C | 2 | 0 |

If A runs much slower than B and C, then A is still the only one to run. 

### With this PR


Assume 5 clients on slave node:

| task | weight | dedicated num client goroutines | % likelihood of execution |
| - | - | - | - |
| A | 45 | 0 | 91.83% |
| B | 2 | 0 | 4.08% |
| C | 2 | 0 | 4.08% |

If A runs slower than B or C, then it will "hog" the results. Let's say task A takes 1 second to execute, and B and C each take 100ms. Then let's look at one goroutine in a 10 second window. The first time it picks up work, it has a 91.83% chance to pick up A, which executes for 1 second. Then it would have a 91.83% chance to pick up A again (and it could do this 10 times). The chances that this one goroutine pick up A 10 times is 91.83%^10 = 42.64%. And it gets more complicated from there.

Essentially boomer cannot _know_ how long a task is going to take to execute and can't dynamically weight the tasks. Correct?


## Next steps
 This PR is attempting to give a percentage chance for each task every time a goroutine picks up work.

What should we do? Should we force consumers to just use `WeighingTaskSet` (does that provide the timing balancing that you're asking about?) or can we move forward with this change or a similar one?