@myzhan 2020-01-17
@joshprzybyszewski-wf Hi, I'm planning to accept this change. Can you avoid creating a random number every time we pick a task? Instead, we create an array and shuffle it. Then we can do round-robin selection to pick a task.