@joshprzybyszewski-wf 2020-01-17
@myzhan I was able to cut down some arithmetic, but I think this is the best way to get a random task (because otherwise I may be allocing a million+ elem slice just to iterate over for a random task)