@myzhan 2020-01-22
@joshprzybyszewski-wf Hi, there is a nil pointer usage in your last commit. I have fixed it and create random sources in getRandomTask, just to avoid usage of the global random source, which has a lock inside.